Add input history, delete, reverse
I fundamentally changed how the calculator operated. I removed the
variables for previous numbers and current operators, and instead
have an array of inputs that is looped through whenever a cacluation is
made. With this, I was able to add the ability for users to delete their
last input as well as reverse the sign of the number.